+++ kwarray/__init__.py
@@ -20,7 +20,7 @@ __explicit__ = [
     'ArrayAPI'
 ]
 
-__version__ = '0.5.4'
+__version__ = '0.5.5'
 
 from kwarray.arrayapi import ArrayAPI