+++ CHANGELOG.md
@@ -3,7 +3,15 @@
 This changelog follows the specifications detailed in: [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html), although we have not yet reached a `1.0.0` release.
 
-## [Version 0.5.4] - Unreleased
+## Version 0.5.5
+
+### Added
+* `DataFrameLight.from_pandas`
+* `DataFrameLight.iterrows`
+* `DataFrameLight.pandas` in favor of `_pandas`
+
+## Version 0.5.4 - Released 2020-02-19 
+
 
 ### Added
 * Add `FlatIndexer`